Refatoração geral

+++ src/main/java/com/pw3/aleatorypost/model/dao/CommentDAO.java
@@ -7,8 +7,8 @@ import com.pw3.aleatorypost.model.domain.Comment;
 public interface CommentDAO {
     void save(Comment comment);
     void update(Comment comment);
-    void delete(Long id);
-    Comment findById(Long id);
-    List<Comment> findByPostId(Long postId);
+    void delete(Integer id);
+    Comment findById(Integer id);
+    List<Comment> findByPostId(Integer postId);
     List<Comment> findAll();
 }
