+++ src/main/java/com/pw3/aleatorypost/model/dao/UserDAO.java
@@ -7,6 +7,6 @@ import com.pw3.aleatorypost.model.domain.User;
 public interface UserDAO {
     void save(User user);
     void update(User task);
-    User findById(Long id);
+    User findById(Integer id);
     User findByEmail(String email);
 }
