+++ src/main/java/com/pw3/aleatorypost/model/dao/UserDAOImpl.java
@@ -8,7 +8,7 @@ import com.pw3.aleatorypost.model.domain.User;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class UserDAOImpl extends AbstractDAO<User, Long> implements UserDAO {
+public class UserDAOImpl extends AbstractDAO<User, Integer> implements UserDAO {
 
     @Override
     public User findByEmail(String email) {
